updated groupchat schema to contain creator email and added new mutation resolver for updating statuses

+++ server/resolvers/GroupChatResolver.ts
@@ -124,7 +124,7 @@ export class GroupChatResolver {
     if (!user) {
       return null;
     }
-    const newGroupChat = await GroupChatModel.create({ ...groupchatInfo });
+    const newGroupChat = await GroupChatModel.create({ ...groupchatInfo, createdBy: email});
     await UserModel.updateOne(
       { email },
       { $push: { groupChatsCreated: newGroupChat._id } }
@@ -156,10 +156,26 @@ export class GroupChatResolver {
     return result;
   }
 
+  @Mutation(() => GroupChat, { nullable: true })
+  async updateStatus(@Arg("id") id:string, @Arg("status") status:string) {
+    const groupChat = await GroupChatModel.findOne({ _id: id });
+		if (!groupChat) {
+			return null;
+		}
+		groupChat.status = status;
+		const result = await groupChat.save();
+		return result;
+  }
+
   @Mutation(returns => Boolean)
   async deleteGroupChat(@Arg("id") id:string) {
-    const result = await GroupChatModel.deleteOne({ _id: id});
-    return result;
+    const chat = await GroupChatModel.findOne({ _id: id});
+    if(chat != undefined) {
+      const result = await UserModel.updateOne({ email: chat.createdBy }, {$pull: {groupChatsCreated: id}});
+      const result2 = await GroupChatModel.deleteOne({ _id: id});
+      return result && (result2.n == 1);
+    }
+    return false;
   }
